git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@6447 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ tools/riddler/java/edu/cmu/sphinx/tools/riddler/types/AudioID.java
@@ -25,19 +25,19 @@ import java.util.List;
  * @see RegionOfAudioID
  */
 public class AudioID {
-    int id;
+    long id;
     List<RegionOfAudioID> audioRegionIDs;
 
-    public AudioID(int id, List<RegionOfAudioID> audioRegionIDs) {
+    public AudioID(long id, List<RegionOfAudioID> audioRegionIDs) {
         this.id = id;
         this.audioRegionIDs = audioRegionIDs;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
